# Fix preformance issues