--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,4 @@
 target
 /classes
+.idea
+*.iml